Fix appointment cancellation and availability deletion. Normalize appointment data to ensure consistent IDs. Add error handling and user feedback for these actions. Also, add an alert dialog to show success or error messages when canceling appointments.

+++ backend/src/main/java/com/medibook/appointment/controllers/AuthController.java
@@ -105,7 +105,7 @@ public class AuthController {
 
                 User User = userService.findUserByUsername(userDetails.getUsername()).get();
                 Integer patientProfileId = User.getPatientProfile() != null ? User.getPatientProfile().getId() : null;
-                Integer doctorProfileId = User.getDoctorProfile() != null ? User.getDoctorProfile().getId() : null;
+                Long doctorProfileId = User.getDoctorProfile() != null ? User.getDoctorProfile().getId() : null;
 
                 return  ResponseEntity.ok(new JwtResponse(jwt,
                         userDetails.getId(),
